backend/rag_indexer/requirements.txt
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.33.1
google-cloud-logging==3.6.0
pypdf==3.16.0
functions-framework==3.4.0
